Clean up GMW namespace and use encapsulation like the other protocols do
