+++ lib/client/api.js
@@ -5,6 +5,7 @@ var crypto_provider = require('./api/crypto_provider.js');
 var synchronization = require('./api/synchronization.js');
 var protocols = require('./api/protocols.js');
 var bitsProtocols = require('./api/bits.js');
+var gmwProtocols = require('./api/gmw.js');
 
 
 
@@ -16,4 +17,5 @@ module.exports = function (jiffClient) {
   synchronization(jiffClient);
   protocols(jiffClient);
   bitsProtocols(jiffClient);
+  gmwProtocols(jiffClient);
 };
