+++ lib/client/api/sharing.js
@@ -2,8 +2,8 @@ var shareProtocol = require('../protocols/shamir/share.js');
 var openProtocol = require('../protocols/shamir/open.js');
 var reshareProtocol = require('../protocols/shamir/reshare.js');
 var arraysSharing = require('../protocols/arrays/api.js');
-var gmwShareProtocol=require('../protocols/gmw_bit/gmw_share.js');
-var gmwOpenProtocol = require('../protocols/gmw_bit/gmw_open.js');
+var gmwShareProtocol=require('../protocols/gmw/share.js');
+var gmwOpenProtocol = require('../protocols/gmw/open.js');
 
 module.exports = function (jiffClient) {
   /**
