+++ lib/client/share.js
@@ -3,7 +3,7 @@ var arithmetic = require('./protocols/numbers/arithmetic.js');
 var comparison = require('./protocols/numbers/comparison.js');
 var protocols = require('./protocols/numbers/protocols.js');
 var booleans = require('./protocols/booleans/boolean.js');
-var gmw = require('./protocols/gmw_bit/gmw.js');
+var gmw = require('./protocols/gmw/gates.js');
 //
 // a metaclass that creates SecretShare classes when given a jiff instance
 // alternatively, we can think of this as a factory for a secret share prototypes/constructors given a jiff instance
