+++ lib/jiff-client.js
@@ -61,10 +61,7 @@ var api = require('./client/api.js');
 var preprocessingMap = require('./client/preprocessing/map.js');
 var preprocessingAPI = require('./client/preprocessing/api.js');
 var preprocessingDaemon = require('./client/preprocessing/daemon.js');
-// used for gmw bit computation( OT lib)
-let IO = require('./client/protocols/gmw_bit/IO.js');
 
-let OT = require('1-out-of-n')(IO);
 /**
  * Creates a new jiff client instance.
  * @class
@@ -333,6 +330,20 @@ function JIFFClient(hostname, computation_id, options) {
   this.protocols.bits = {};
 
   /**
+   * A collection of useful protocols for manipulating GMW shares.
+   * @member {!gmw} gmw
+   * @memberof protocols
+   */
+  this.protocols.gmw = {};
+
+  /**
+   * A collection of useful protocols for manipulating GMW bit arrays, and transforming them from and to regular numeric shares.
+   * @member {!bits} bits
+   * @memberof protocols.gmw
+   */
+  this.protocols.gmw.bits = {};
+
+  /**
    * Stores pre-computed values (beaver triples, random bits, etc) used to aid/speed up the main processes.
    * @type {!object}
    */
@@ -423,7 +434,7 @@ function JIFFClient(hostname, computation_id, options) {
    * Stores the OT for gmw bit computation.
    * @type {Promise}
    */
-  this.OT=OT;
+  this.OT = {};
 
   // Add user facing API
   api(this);
